+++ be/test/service/service_be/internal_service_test.cpp
@@ -20,6 +20,7 @@
 #include "common/utils.h"
 #include "exec/tablet_sink_index_channel.h"
 #include "runtime/exec_env.h"
+#include "service/brpc_service_test_util.h"
 
 namespace starrocks {
 
@@ -33,19 +34,6 @@ TEST_F(InternalServiceTest, test_get_info_timeout_invalid) {
     auto st = Status(response.status());
     ASSERT_TRUE(st.is_time_out());
 }
-
-class MockClosure : public ::google::protobuf::Closure {
-public:
-    MockClosure() = default;
-    ~MockClosure() override = default;
-
-    void Run() override { _run.store(true); }
-
-    bool has_run() { return _run.load(); }
-
-private:
-    std::atomic_bool _run = false;
-};
 
 TEST_F(InternalServiceTest, test_tablet_writer_add_chunks_via_http) {
     BackendInternalServiceImpl<PInternalService> service(ExecEnv::GetInstance());
